# Make one of the slowest io.fits tests faster.