--- a/astropy/io/fits/tests/test_image.py
+++ b/astropy/io/fits/tests/test_image.py
@@ -1506,7 +1506,8 @@ class TestCompressedImage(FitsTestCase):
     def test_lossless_gzip_compression(self):
         """Regression test for https://aeon.stsci.edu/ssb/trac/pyfits/ticket/198"""
 
-        noise = np.random.normal(size=(1000, 1000))
+        rng = np.random.RandomState(seed=42)
+        noise = rng.normal(size=(20, 20))
 
         chdu1 = fits.CompImageHDU(data=noise, compression_type='GZIP_1')
         # First make a test image with lossy compression and make sure it