Generate fewer random values in test.

EDIT:
Original: [`1.16s call   astropy/io/fits/tests/test_image.py::TestCompressedImage::test_lossless_gzip_compression`](https://travis-ci.org/astropy/astropy/jobs/586576960#L2134)

With changes: [`0.56s call     astropy/io/fits/tests/test_image.py::TestCompressedImage::test_lossless_gzip_compression`](https://travis-ci.org/astropy/astropy/jobs/586857021#L2153)

Locally the speedup was more pronounced 3s -> 0.4s.