@bsipocz 2019-09-18
Yes, I started to see the coordinates failure locally yesterday, not yet sure about the reason (well, most probably is the new doctestplus release, but I'm not certain about it yet).