@bsipocz 2019-09-18
Hmm, interesting. It's indeed `pytest-doctestplus` related, but it doesn't seem to break it rather than actually runs it.
With pytest-doctestplus==0.3.0, nothing is collected:
```
$ ~/munka/devel/astropy [hacking|⚑ 20] $ pt -t astropy/coordinates/distances.py
running test
installing to temporary directory: /private/var/folders/dc/hsm7tqpx2d57n7vb3k1l81xw0000gq/T/astropy-test-db30ymsh
warning: no files found matching '*' under directory 'scripts'
no previously-included directories found matching 'docs/_build'
no previously-included directories found matching 'build'
warning: no files found matching '*.pyx' under directory 'astropy_helpers/astropy_helpers'
warning: no files found matching '*.h' under directory 'astropy_helpers/astropy_helpers'
warning: no files found matching '*.rst' under directory 'astropy_helpers/astropy_helpers'
no previously-included directories found matching 'astropy_helpers/build'
warning: no previously-included files matching '*.o' found anywhere in distribution
/usr/local/lib/python3.7/site-packages/pytest_doctestplus/plugin.py:10: DeprecationWarning: the imp module is deprecated in favour of importlib; see the module's documentation for alternative uses
  import imp
================================================= test session starts ==================================================
platform darwin -- Python 3.7.4, pytest-5.1.2, py-1.8.0, pluggy-0.12.0

Running tests with Astropy version 4.0.dev25737.
Running tests in astropy/coordinates/distances.py.

Date: 2019-09-18T13:03:07

Platform: Darwin-17.7.0-x86_64-i386-64bit

Executable: /usr/local/opt/python/bin/python3.7

Full Python Version: 
3.7.4 (default, Jul  9 2019, 18:15:00) 
[Clang 10.0.0 (clang-1000.11.45.5)]

encodings: sys: utf-8, locale: UTF-8, filesystem: utf-8
byteorder: little
float info: dig: 15, mant_dig: 15

Numpy: 1.18.0.dev0+4d13d4c
Scipy: 1.3.1
Matplotlib: 3.1.1.post2162+g770d59d42
h5py: 2.8.0
Pandas: 0.24.2
Cython: 0.29.13
Scikit-image: 0.15.0
asdf: not available
astropy_helpers: 3.1rc1
Using Astropy options: remote_data: none.

rootdir: /private/var/folders/dc/hsm7tqpx2d57n7vb3k1l81xw0000gq/T/astropy-test-db30ymsh/lib/python3.7/site-packages, inifile: setup.cfg
plugins: forked-1.0.2, arraydiff-0.3, cov-2.6.0, xonsh-0.8.3, dependency-0.4.0, doctestplus-0.3.0, requests-mock-1.6.0, openfiles-0.4.0.dev0, remotedata-0.3.2.dev0
collected 0 items                                                                                                      

================================================ no tests ran in 2.06s =================================================
```

While with 0.4.0:
```
================================================= test session starts ==================================================
platform darwin -- Python 3.7.4, pytest-5.1.2, py-1.8.0, pluggy-0.12.0

Running tests with Astropy version 4.0.dev25737.
Running tests in astropy/coordinates/distances.py.

Date: 2019-09-18T13:03:44

Platform: Darwin-17.7.0-x86_64-i386-64bit

Executable: /usr/local/opt/python/bin/python3.7

Full Python Version: 
3.7.4 (default, Jul  9 2019, 18:15:00) 
[Clang 10.0.0 (clang-1000.11.45.5)]

encodings: sys: utf-8, locale: UTF-8, filesystem: utf-8
byteorder: little
float info: dig: 15, mant_dig: 15

Numpy: 1.18.0.dev0+4d13d4c
Scipy: 1.3.1
Matplotlib: 3.1.1.post2162+g770d59d42
h5py: 2.8.0
Pandas: 0.24.2
Cython: 0.29.13
Scikit-image: 0.15.0
asdf: not available
astropy_helpers: 3.1rc1
Using Astropy options: remote_data: none.

rootdir: /private/var/folders/dc/hsm7tqpx2d57n7vb3k1l81xw0000gq/T/astropy-test-mhk6rxyb/lib/python3.7/site-packages, inifile: setup.cfg
plugins: forked-1.0.2, arraydiff-0.3, cov-2.6.0, xonsh-0.8.3, dependency-0.4.0, doctestplus-0.4.0, requests-mock-1.6.0, openfiles-0.4.0.dev0, remotedata-0.3.2.dev0
collected 1 item                                                                                                       

astropy/coordinates/distances.py F                                                                               [100%]

======================================================= FAILURES =======================================================
___________________________________ [doctest] astropy.coordinates.distances.Distance ___________________________________
075         If ``z`` is provided with a ``unit`` or ``cosmology`` is provided
076         when ``z`` is *not* given, or ``value`` is given as well as ``z``.
077 
078 
079     Examples
080     --------
081     >>> from astropy import units as u
082     >>> from astropy import cosmology
083     >>> from astropy.cosmology import WMAP5, WMAP7
084     >>> cosmology.set_current(WMAP7)
UNEXPECTED EXCEPTION: AttributeError("module 'astropy.cosmology' has no attribute 'set_current'")
Traceback (most recent call last):

  File "/usr/local/Cellar/python/3.7.4/Frameworks/Python.framework/Versions/3.7/lib/python3.7/doctest.py", line 1329, in __run
    compileflags, 1), test.globs)

  File "<doctest astropy.coordinates.distances.Distance[3]>", line 1, in <module>

AttributeError: module 'astropy.cosmology' has no attribute 'set_current'

/private/var/folders/dc/hsm7tqpx2d57n7vb3k1l81xw0000gq/T/astropy-test-mhk6rxyb/lib/python3.7/site-packages/astropy/coordinates/distances.py:84: UnexpectedException
================================================== 1 failed in 2.57s ===================================================
```