@MSeifert04 2019-09-18
@bsipocz I can fix a few of these, but probably won't have the time to fix them all.