@MSeifert04 2019-09-19
Coverage decrease is real. It seems like we don't have any other compressed file exceeding the IO_BUFFER_SIZE...