@MSeifert04 2019-09-19
I think it's not worth it fixing the coverage failure. It's just the "false" case of a condition which has no "else".

The only way to reliably cover this would be to use an (highly artifical and maybe depending on implementation details) image which happens to have the matching `max_elem`. 

Or to rewrite the logic to be more general (which would trivially be covered for any large-enough compressed image):

```
    if (*bufsize <= IOBUFLEN) {
        *bufsize = IOBUFLEN;
    } else {
        *bufsize = ((*bufsize - 1) / IOBUFLEN + 1) * IOBUFLEN;
    }
```