@saimn 2019-09-20
@MSeifert04 - Maybe a small increase of the array size would be enough to have more than on FITS block ?