@MSeifert04 2019-09-20
> Maybe a small increase of the array size would be enough to have more than on FITS block ?

Problem is that it needs to be exactly a complete multiple of the blocksize to "fix the coverage". Just having more than one block isn't enough (the content of `else if` is covered).