@saimn 2019-09-20
> the content of else if is covered

Ah ok, I thought that it was decreasing because of this. 

We probably do not need to worry too much about this then. Thanks for the PR!